moving to vulkan
decided that its best to move the debugger to vulkan. this is mainly for Linux people, as the official samples for jolt are D3D12 Only, that includes the debugger. so we might as well deal with the problem early, and not have to worry later, when our framework is already developed.